### 2022-10-09 #4837 @editorialbot
[PRE REVIEW]: OpenMSIStream: A Python package for facilitating integration of streaming data in diverse laboratory environments
**Submitting author:** @eminizer (<a href="http://orcid.org/0000-0003-4591-2225">Margaret Eminizer</a>)
**Repository:** https://github.com/openmsi/openmsistream
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
