### 2022-10-14 3e18844
some bugfixes for working with Windows Services
more verbose NSSM install, fixed a bug where environment variables would be repeatedly set, and fix problems caused by spaces in paths on Windows
