+++ openmsistream/services/windows_service_manager.py
@@ -29,8 +29,7 @@ class WindowsServiceManager(ServiceManagerBase) :
 
     @property
     def env_var_names(self):
-        for evn in super().env_var_names :
-            yield evn
+        env_var_names = super().env_var_names
         #get the names of environment variables from the env_var file
         if self.env_var_filepath.is_file() :
             with open(self.env_var_filepath,'r') as fp :
@@ -38,7 +37,9 @@ class WindowsServiceManager(ServiceManagerBase) :
             for line in lines :
                 linestrip = line.strip()
                 if linestrip!='' :
-                    yield linestrip
+                    env_var_names.add(linestrip)
+        for evn in env_var_names :
+            yield evn
 
     def install_service(self) :
         """
@@ -50,11 +51,12 @@ class WindowsServiceManager(ServiceManagerBase) :
         #find or install NSSM to run the executable
         self.__find_install_nssm()
         #run NSSM to install the service
-        cmd = f'{SERVICE_CONST.NSSM_PATH} install {self.service_name} \"{sys.executable}\" \"{self.exec_fp}\"'
-        run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
-        run_cmd_in_subprocess(['powershell.exe',
-                               f'{SERVICE_CONST.NSSM_PATH} set {self.service_name} DisplayName {self.service_name}'],
-                               logger=self.logger)
+        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','install',f'{self.service_name}',
+                f'\"{sys.executable}\"',f'\"{self.exec_fp}\"']
+        run_cmd_in_subprocess(cmds,logger=self.logger)
+        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','set',f'{self.service_name}',
+                'DisplayName',f'{self.service_name}']
+        run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'Done installing {self.service_name}')
 
     def start_service(self) :
@@ -63,8 +65,8 @@ class WindowsServiceManager(ServiceManagerBase) :
         """
         self.logger.info(f'Starting {self.service_name}...')
         #start the service using net
-        cmd = f'net start {self.service_name}'
-        run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
+        cmds = ['powershell.exe','net','start',f'{self.service_name}']
+        run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'Done starting {self.service_name}')
 
     def service_status(self) :
@@ -74,8 +76,8 @@ class WindowsServiceManager(ServiceManagerBase) :
         #find or install NSSM in the current directory
         self.__find_install_nssm()
         #get the service status
-        cmd = f'{SERVICE_CONST.NSSM_PATH} status {self.service_name}'
-        result = run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
+        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','status',f'{self.service_name}']
+        result = run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'{self.service_name} status: {result.decode()}')
 
     def stop_service(self) :
@@ -84,8 +86,8 @@ class WindowsServiceManager(ServiceManagerBase) :
         """
[…]
