### 2022-10-14 2f9ae3d
forgot that powershell commands have to be invoked differently in subprocesses
