+++ openmsistream/services/windows_service_manager.py
@@ -51,12 +51,10 @@ class WindowsServiceManager(ServiceManagerBase) :
         #find or install NSSM to run the executable
         self.__find_install_nssm()
         #run NSSM to install the service
-        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','install',f'{self.service_name}',
-                f'\"{sys.executable}\"',f'\"{self.exec_fp}\"']
-        run_cmd_in_subprocess(cmds,logger=self.logger)
-        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','set',f'{self.service_name}',
-                'DisplayName',f'{self.service_name}']
-        run_cmd_in_subprocess(cmds,logger=self.logger)
+        cmd = f'\"{SERVICE_CONST.NSSM_PATH}\" install {self.service_name} \"{sys.executable}\" \"{self.exec_fp}\"'
+        run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
+        cmd = f'\"{SERVICE_CONST.NSSM_PATH}\" set {self.service_name} DisplayName {self.service_name}'
+        run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
         self.logger.info(f'Done installing {self.service_name}')
 
     def start_service(self) :
@@ -65,7 +63,7 @@ class WindowsServiceManager(ServiceManagerBase) :
         """
         self.logger.info(f'Starting {self.service_name}...')
         #start the service using net
-        cmds = ['powershell.exe','net','start',f'{self.service_name}']
+        cmds = ['powershell.exe',f'net start {self.service_name}']
         run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'Done starting {self.service_name}')
 
@@ -76,7 +74,7 @@ class WindowsServiceManager(ServiceManagerBase) :
         #find or install NSSM in the current directory
         self.__find_install_nssm()
         #get the service status
-        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','status',f'{self.service_name}']
+        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\" status {self.service_name}']
         result = run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'{self.service_name} status: {result.decode()}')
 
@@ -86,7 +84,7 @@ class WindowsServiceManager(ServiceManagerBase) :
         """
         self.logger.info(f'Stopping {self.service_name}...')
         #stop the service using net
-        cmds = ['powershell.exe','net','stop',f'{self.service_name}']
+        cmds = ['powershell.exe',f'net stop {self.service_name}']
         run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info(f'Done stopping {self.service_name}')
 
@@ -106,14 +104,14 @@ class WindowsServiceManager(ServiceManagerBase) :
         #find or install NSSM in the current directory
         self.__find_install_nssm()
         #using NSSM
-        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\"','remove',f'{self.service_name}','confirm']
+        cmds = ['powershell.exe',f'\"{SERVICE_CONST.NSSM_PATH}\" remove {self.service_name} confirm']
         run_cmd_in_subprocess(cmds,logger=self.logger)
         self.logger.info('Service removed')
         #remove NSSM if requested
         if remove_nssm :
[…]
