### 2022-10-14 ac7c4db
cd into the working directory to run NSSM commands
