### 2022-10-14 1c032fb
no quotes around the name of the NSSM executable
