### 2022-10-14 782cab9
run nssm with "./nssm.exe"
