### 2022-10-14 87c3546
Adjust path to Python executable for paths that have spaces in parts of them
