### 2022-10-14 c468e9b
add a slash to the first part of the executable path if it ends with a colon
like a drive specification would be, otherwise it doesn't get added.
