+++ openmsistream/services/windows_service_manager.py
@@ -54,7 +54,10 @@ class WindowsServiceManager(ServiceManagerBase) :
         #run NSSM to install the service
         with change_dir(SERVICE_CONST.NSSM_PATH.parent) :
             cmd = f'./{SERVICE_CONST.NSSM_PATH.name} install {self.service_name} "{sys.executable}" '
-            cmd+= f'"\""""{self.exec_fp}\"""'
+            if ' ' in str(self.exec_fp) :
+                cmd+= f'"\""""{self.exec_fp}\"""'
+            else :
+                cmd+= f'"{self.exec_fp}"'
             run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
             cmd = f'./{SERVICE_CONST.NSSM_PATH.name} set {self.service_name} DisplayName {self.service_name}'
             run_cmd_in_subprocess(['powershell.exe',cmd],logger=self.logger)
