### 2022-10-14 978aad2
add more DLLs, and import OpenMSIStream on its own
