+++ openmsistream/services/linux_service_manager.py
@@ -158,11 +158,12 @@ class LinuxServiceManager(ServiceManagerBase) :
         code = f'''\
             [Unit]
             Description = {description}
-            Requires = network-online.target remote-fs.target
-            After = network-online.target remote-fs.target
+            Requires = network-online.target local-fs.target remote-fs.target
+            After = network-online.target local-fs.target remote-fs.target
 
             [Service]
             Type = simple
+            KillMode = none
             User = {os.path.expandvars('$USER')}
             ExecStart = {sys.executable} {self.exec_fp}'''
         if self.env_vars_needed :
@@ -170,8 +171,8 @@ class LinuxServiceManager(ServiceManagerBase) :
             EnvironmentFile = {self.env_var_filepath}'''
         code+=f'''\n\
             WorkingDirectory = {pathlib.Path().resolve()}
-            Restart = on-failure
-            RestartSec = 30
+            Restart = always
+            RestartSec = 120
 
             [Install]
             WantedBy = multi-user.target'''
