### 2022-10-25 0101ebe
split out dataclass table functionality to create append-only version
