+++ openmsistream/utilities/dataclass_table.py
@@ -1,10 +1,11 @@
 """
-A utility class to represent a set of (relatively arbitrary) dataclass objects serialized to/deseralized from
-a corresponding CSV file in an atomic and (relatively) thread-safe way
+A couple utility classes (and a helper function) to represent a set of (relatively arbitrary) dataclass objects
+serialized to/deseralized from a corresponding CSV file in an atomic and (relatively) thread-safe way
 """
 
 #imports
 import pathlib, functools, datetime, typing, copy, os
+from abc import ABC
 from threading import Lock
 from dataclasses import fields, is_dataclass
 import methodtools
@@ -23,9 +24,9 @@ def get_nested_types() :
             nested_types_dict[c_t[0][s_t]] = (c_t[1],s_t,c_t[2])
     return nested_types_dict
 
-class DataclassTable(LogOwner) :
+class DataclassTableBase(LogOwner,ABC) :
     """
-    A class to work with an atomic csv file that's holding dataclass entries in a thread-safe way
+    A base class for DataclassTable objects
 
     :param dataclass_type: The :class:`dataclasses.dataclass` defining the entries in the table/csv file
     :type dataclass_type: :class:`dataclasses.dataclass`
@@ -38,7 +39,6 @@ class DataclassTable(LogOwner) :
 
     DELIMETER = ';' #can't use a comma or containers would display incorrectly
     DATETIME_FORMAT = '%a %b %d, %Y at %H:%M:%S'
-    THREAD_LOCK = Lock()
     UPDATE_FILE_EVERY = 5 #only update the .csv file automatically every 5 seconds to make updates less expensive
 
     NESTED_TYPES = get_nested_types()
@@ -52,9 +52,9 @@ class DataclassTable(LogOwner) :
         header_line = ''
         #add an extra line when running on Windows to seamlessly open the file in Excel
         if os.name=='nt' :
-            header_line+=f'sep={DataclassTable.DELIMETER}\n'
+            header_line+=f'sep={self.DELIMETER}\n'
         for fieldname in self.__field_names :
-            header_line+=f'{fieldname}{DataclassTable.DELIMETER}'
+            header_line+=f'{fieldname}{self.DELIMETER}'
         return header_line[:-1]
 
     @property
@@ -62,7 +62,7 @@ class DataclassTable(LogOwner) :
         """
         All recognized object hex addresses
         """
-        return list(self.__entry_objs.keys())
+        return list(self._entry_objs.keys())
 
     @property
     def filepath(self) :
@@ -70,13 +70,6 @@ class DataclassTable(LogOwner) :
         Path to the CSV file
         """
         return self.__filepath
-
-    @property
-    def lock(self) :
-        """
-        A thread lock to use for ensuring only one thread is interacting with the :class:`~DataclassTable` at a time
-        """
-        return DataclassTable.THREAD_LOCK
 
     #################### PUBLIC FUNCTIONS ####################
 
@@ -86,6 +79,8 @@ class DataclassTable(LogOwner) :
         """
         #init the LogOwner
         super().__init__(**kwargs)
+        #create the thread lock for the table
+        self.lock = Lock()
         #figure out what type of objects the table/file will be describing
         self.__dataclass_type = dataclass_type
         if not is_dataclass(self.__dataclass_type) :
@@ -99,10 +94,10 @@ class DataclassTable(LogOwner) :
         #figure out where the csv file should go
         self.__filepath = filepath if filepath is not None else pathlib.Path() / f'{self.__dataclass_type.__name__}.csv'
         #set some other variables
-        self.__entry_objs = {}
-        self.__entry_lines = {}
+        self._entry_objs = {}
+        self._entry_lines = {}
         #read or create the file to finish setting up the table
-        self.__file_last_updated = datetime.datetime.now()
[…]
