+++ openmsistream/kafka_wrapper/utilities.py
@@ -24,7 +24,7 @@ def add_kwargs_to_configs(configs,logger,**kwargs) :
         new_configs[config_name]=arg
     return new_configs
 
-def default_producer_callback(err,msg,logger=None,**other_kwargs) :
+def default_producer_callback(err,msg,prodid,logger=None,**other_kwargs) :
     """
     Default callback function to use for testing whether a message has been successfully produced to the topic
     """
@@ -32,15 +32,15 @@ def default_producer_callback(err,msg,logger=None,**other_kwargs) :
         err = msg.error()
     if err is not None: #raise an error if the message wasn't sent successfully
         if err.fatal() :
-            logmsg =f'ERROR: fatally failed to deliver message with kwargs "{other_kwargs}". '
-            logmsg+=f'MESSAGE DROPPED. Error reason: {err.str()}'
+            logmsg =f'ERROR: Producer with ID {prodid} fatally failed to deliver message with kwargs '
+            logmsg+=f'"{other_kwargs}". MESSAGE DROPPED. Error reason: {err.str()}'
             if logger is not None :
                 logger.error(logmsg)
             else :
                 raise RuntimeError(logmsg)
         elif not err.retriable() :
-            logmsg =f'ERROR: Failed to deliver message with kwargs "{other_kwargs}" and cannot retry. '
-            logmsg+= f'MESSAGE DROPPED. Error reason: {err.str()}'
+            logmsg =f'ERROR: Producer with ID {prodid} failed to deliver message with kwargs "{other_kwargs}" '
+            logmsg+= f'and cannot retry. MESSAGE DROPPED. Error reason: {err.str()}'
             if logger is not None :
                 logger.error(logmsg)
             else :
