+++ openmsistream/data_file_io/actor/file_registry/producer_file_registry.py
@@ -7,7 +7,7 @@ in the process of being produced/have been fully produced
 import pathlib, datetime
 from typing import Set
 from dataclasses import dataclass
-from ....utilities import DataclassTable, LogOwner
+from ....utilities import DataclassTableReadOnly, DataclassTableAppendOnly, DataclassTable, LogOwner
 
 @dataclass
 class RegistryLineInProgress :
@@ -40,21 +40,67 @@ class ProducerFileRegistry(LogOwner) :
     of which files have been uploaded to a particular topic
     """
 
-    def __init__(self,dirpath,topic_name,*args,**kwargs) :
+    def __init__(self,dirpath,topic_name,*args,max_completed_lines_per_file=1000,**kwargs) :
         """
         dirpath    = path to the directory that should contain the csv file
         topic_name = the name of the topic that will be produced to (used in the filename)
         """
         super().__init__(*args,**kwargs)
         #A table to list the files currently in progress
-        in_prog_filepath = dirpath / f'files_to_upload_to_{topic_name}.csv'
+        in_prog_filepath = dirpath / f'upload_to_{topic_name}_in_progress.csv'
         self.__in_prog = DataclassTable(dataclass_type=RegistryLineInProgress,filepath=in_prog_filepath,
                                         logger=self.logger)
-        #A table to list the files that have been completely produced
-        completed_filepath = dirpath / f'files_fully_uploaded_to_{topic_name}.csv'
-        self.__completed = DataclassTable(dataclass_type=RegistryLineCompleted,filepath=completed_filepath,
-                                          logger=self.logger)
+        #Pattern for the "completed" .csv file names
+        self.completed_filepath_pattern = dirpath / f'uploaded_to_{topic_name}.csv'
+        #Consolidate existing "completed" csv files into one
+        self.consolidate_completed_files()
+        #Make a dictionary to hold tables of any newly received "completed" entries
+        self.__completed_tables_by_path = {}
+        #keep track of the maximum number of lines per "completed" file
+        self.max_completed_lines_per_file = max_completed_lines_per_file
 
+    def __del__(self) :
+        self.consolidate_completed_files()
+
+    def consolidate_completed_files(self) :
+        """
+        Search the directory holding the files for all "completed" file entries and write them into a single file
+        """
+        #add any lines from files in the directory to the one consolidated file at the expected path
+        globpattern = f'{self.completed_filepath_pattern.stem}*{self.completed_filepath_pattern.suffix}'
+        consolidated_file = None
+        for fp in self.completed_filepath_pattern.parent.glob(globpattern) :
+            if fp==self.completed_filepath_pattern :
+                continue
+            if fp in self.__completed_tables_by_path :
+                self.__completed_tables_by_path[fp].dump_to_file()
+            file_to_add = DataclassTableReadOnly(dataclass_type=RegistryLineCompleted,filepath=fp,logger=self.logger)
+            if consolidated_file is None :
+                consolidated_file = DataclassTableAppendOnly(dataclass_type=RegistryLineCompleted,
+                                                             filepath=self.completed_filepath_pattern,
+                                                             logger=self.logger)
+            consolidated_file.add_entries(file_to_add.objects)
+        if consolidated_file is None :
+            return
+        #dump out the consolidated file
+        consolidated_file.dump_to_file()
+        #make sure that all the lines from the individual files have been successfully copied
+        all_lines = (DataclassTableReadOnly(dataclass_type=RegistryLineCompleted,
+                                            filepath=self.completed_filepath_pattern,logger=self.logger)).lines
+        for fp in self.completed_filepath_pattern.parent.glob(globpattern) :
+            if fp==self.completed_filepath_pattern :
+                continue
[…]
