+++ openmsistream/utilities/__init__.py
@@ -1,10 +1,12 @@
 """Helper classes/wrappers/infrastructure used throughout OpenMSIStream"""
 
 from .logging import Logger, LogOwner
-from .dataclass_table import DataclassTable
+from .dataclass_table import DataclassTableReadOnly, DataclassTableAppendOnly, DataclassTable
 
 __all__ = [
     'Logger',
     'LogOwner',
+    'DataclassTableReadOnly',
+    'DataclassTableAppendOnly',
     'DataclassTable',
 ]

[…]
