### 2022-10-25 11449d6
Updates for producer logs of fully uploaded files
Each thread gets its own file, capped at 1,000 entries each, to prevent the atomic update I/O dominating the runtime each time a file is found. Also some logging updates.
