### 2022-10-25 6f26e5d
some bugfixes for CI tests
