### 2022-10-26 97a8cc3
fix multithreading for DataFileUploadDirectory files in progress table
and some more logging adjustments
