+++ openmsistream/data_file_io/actor/file_registry/producer_file_registry.py
@@ -98,10 +98,8 @@ class ProducerFileRegistry(LogOwner) :
                     errmsg+= 'Duplicate entries may be present.'
                     raise RuntimeError(errmsg)
             del added_file
-            time.sleep(1)
             if fp in self.__completed_tables_by_path :
                 self.__completed_tables_by_path.pop(fp)
-                time.sleep(1)
             fp.unlink()
 
     def get_incomplete_filepaths_and_chunks(self) :
@@ -128,6 +126,8 @@ class ProducerFileRegistry(LogOwner) :
         Register a chunk as having been successfully produced by a particular producer
         Returns True if all chunks for the file have been produced
         """
+        #acquire the lock so no other threads can change the in progress file
+        self.__in_prog.lock.acquire()
         #get a dictionary of the existing object addresses keyed by their filepaths
         existing_obj_addresses = self.__in_prog.obj_addresses_by_key_attr('filepath')
         #if the file is already recognized as in progress
@@ -147,6 +147,7 @@ class ProducerFileRegistry(LogOwner) :
         if len(existing_obj_addresses[filepath])!=1 :
             errmsg = f'ERROR: found {len(existing_obj_addresses[filepath])} files in the producer registry '
             errmsg+= f'for filepath {filepath}'
+            self.__in_prog.lock.release()
             self.logger.error(errmsg,RuntimeError)
         existing_addr = existing_obj_addresses[filepath][0]
         #make sure the total numbers of chunks match
@@ -155,27 +156,27 @@ class ProducerFileRegistry(LogOwner) :
             errmsg = f'ERROR: {filepath} in {self.__class__.__name__} is listed as having '
             errmsg+= f'{existing_n_chunks} total chunks, but the producer callback for this file '
             errmsg+= f'lists {n_total_chunks} chunks! Did the chunk size change?'
+            self.__in_prog.lock.release()
             self.logger.error(errmsg,RuntimeError)
-        #editing the object's attributes has to be done in one thread at a time
-        with self.__in_prog.lock :
-            #get its current state
-            attrs = self.__in_prog.get_entry_attrs(existing_addr,'chunks_delivered','chunks_to_send')
-            #if the chunk is already registered, just return
-            if chunk_i in attrs['chunks_delivered'] :
-                return False
-            #otherwise add/remove it from the sets
-            attrs['chunks_delivered'].add(chunk_i)
-            try :
-                attrs['chunks_to_send'].remove(chunk_i)
-            except KeyError :
-                pass
-            #reset the attributes for the object
-            new_attrs = {
-                'n_chunks_delivered':len(attrs['chunks_delivered']),
-                'n_chunks_to_send':len(attrs['chunks_to_send']),
-                'chunks_delivered':attrs['chunks_delivered'],
-                'chunks_to_send':attrs['chunks_to_send'],
-                }
+        #get its current state
+        attrs = self.__in_prog.get_entry_attrs(existing_addr,'chunks_delivered','chunks_to_send')
+        #if the chunk is already registered, just return
+        if chunk_i in attrs['chunks_delivered'] :
+            self.__in_prog.lock.release()
+            return False
+        #otherwise add/remove it from the sets
+        attrs['chunks_delivered'].add(chunk_i)
+        try :
+            attrs['chunks_to_send'].remove(chunk_i)
+        except KeyError :
+            pass
+        #reset the attributes for the object
+        new_attrs = {
+            'n_chunks_delivered':len(attrs['chunks_delivered']),
+            'n_chunks_to_send':len(attrs['chunks_to_send']),
+            'chunks_delivered':attrs['chunks_delivered'],
+            'chunks_to_send':attrs['chunks_to_send'],
+            }
[…]
