### 2022-10-26 061c5cf
Updates for stream processor/reproducer logs of successfully handled files
in the same vein as the DataFileUploadDirectory edits
