+++ openmsistream/data_file_io/actor/file_registry/stream_handler_registries.py
@@ -4,10 +4,10 @@ being reconstructed from the topic and files that have been fully handled
 """
 
 #imports
-import datetime, re
+import datetime, re, threading
 from abc import ABC
 from dataclasses import dataclass
-from ....utilities import DataclassTable, LogOwner
+from ....utilities import DataclassTableReadOnly, DataclassTableAppendOnly, DataclassTable, LogOwner
 from ...utilities import get_message_prepend
 
 @dataclass
@@ -44,16 +44,18 @@ class StreamHandlerRegistry(LogOwner,ABC) :
     @property
     def in_progress_table(self) :
         """
-        The table listing files in progress
+        A read-only version of the table listing files in progress
         """
-        return self._in_progress_table
+        return self._in_progress_table.as_read_only()
 
     @property
     def succeeded_table(self) :
         """
-        The table listing files that have been successfully handled
+        A read-only version of the table listing files that have been successfully handled
         """
-        return self._succeeded_table
+        self.consolidate_succeeded_files()
+        return DataclassTableReadOnly(StreamHandlerRegistryLineSucceeded,
+                                      filepath=self.succeeded_filepath,logger=self.logger)
 
     @property
     def filepaths_to_rerun(self) :
@@ -91,7 +93,7 @@ class StreamHandlerRegistry(LogOwner,ABC) :
         """
         return len(self.filepaths_to_rerun)
 
-    def __init__(self,in_progress_filepath,succeeded_filepath,*args,**kwargs) :
+    def __init__(self,in_progress_filepath,succeeded_filepath,*args,max_succeeded_lines_per_file=1000,**kwargs) :
         """
         in_progress_filepath = path to the file that should hold the "in_progress" datatable
         succeeded_filepath = path to the file that should hole the "succeeded" datatable
@@ -99,8 +101,10 @@ class StreamHandlerRegistry(LogOwner,ABC) :
         super().__init__(*args,**kwargs)
         self._in_progress_table = DataclassTable(dataclass_type=StreamHandlerRegistryLineInProgress,
                                                   filepath=in_progress_filepath,logger=self.logger)
-        self._succeeded_table = DataclassTable(dataclass_type=StreamHandlerRegistryLineSucceeded,
-                                                  filepath=succeeded_filepath,logger=self.logger)
+        self.succeeded_filepath = succeeded_filepath
+        self.max_succeeded_lines_per_file = max_succeeded_lines_per_file
+        self.__succeeded_tables_by_id = {}
+        self.consolidate_succeeded_files()
 
     def register_file_in_progress(self,dfc) :
         """
@@ -114,22 +118,66 @@ class StreamHandlerRegistry(LogOwner,ABC) :
         """
         self._add_or_modify_in_progress_entry(dfc,self.MISMATCHED_HASH)
 
+    def consolidate_succeeded_files(self) :
+        """
+        Search the directory holding the files for all "completed" file entries and write them into a single file
+        """
+        #add any lines from files in the directory to the one consolidated file at the expected path
+        globpattern = f'{self.succeeded_filepath.stem}*{self.succeeded_filepath.suffix}'
+        consolidated_file = None
+        for fp in self.succeeded_filepath.parent.glob(globpattern) :
+            if fp==self.succeeded_filepath :
+                continue
+            for table in self.__succeeded_tables_by_id.values() :
+                if fp==table.filepath :
+                    table.dump_to_file()
+            file_to_add = DataclassTableReadOnly(dataclass_type=StreamHandlerRegistryLineSucceeded,
+                                                 filepath=fp,logger=self.logger)
+            if consolidated_file is None :
+                consolidated_file = DataclassTableAppendOnly(dataclass_type=StreamHandlerRegistryLineSucceeded,
[…]
