### 2022-10-26 7e85936
CI tests accommodate read-only DataclassTable files
