+++ test/unittests/test_data_file_stream_processor.py
@@ -196,12 +196,12 @@ class TestDataFileStreamProcessor(unittest.TestCase) :
                                         consumer_group_id=CONSUMER_GROUP_ID,
                                         logger=LOGGER)
             self.assertEqual(len(spr.filepaths_to_rerun),1)
-            in_prog_entries = spr.in_progress_table.obj_addresses_by_key_attr('status')
-            succeeded_entries = spr.succeeded_table.obj_addresses
+            in_prog_table = spr.in_progress_table
+            in_prog_entries = in_prog_table.obj_addresses_by_key_attr('status')
+            succeeded_table = spr.succeeded_table
+            succeeded_entries = succeeded_table.obj_addresses
             self.assertEqual(len(succeeded_entries),1)
             self.assertEqual(len(in_prog_entries[spr.FAILED]),1)
-            #get the attributes of the succeeded file to make sure the entry doesn't change
-            succeeded_entry_attrs = spr.succeeded_table.get_entry_attrs(succeeded_entries[0])
         except Exception as e :
             raise e
         finally :
@@ -279,14 +279,11 @@ class TestDataFileStreamProcessor(unittest.TestCase) :
                                         topic_name=TOPIC_NAME,
                                         consumer_group_id=CONSUMER_GROUP_ID,
                                         logger=LOGGER)
-            succeeded_entries = spr.succeeded_table.obj_addresses
+            succeeded_table = spr.succeeded_table
+            succeeded_entries = succeeded_table.obj_addresses
             self.assertTrue(len(succeeded_entries)>=3) #>3 if the topic has files from previous runs in it
-            #get the attributes of the originally succeeded file to make sure the entry hasn't changed
-            succeeded_entries = spr.succeeded_table.obj_addresses_by_key_attr('filename')
+            succeeded_entries = succeeded_table.obj_addresses_by_key_attr('filename')
             self.assertTrue(len(succeeded_entries[TEST_CONST.TEST_DATA_FILE_2_NAME])>=1)
-            testing_entry_attrs = spr.succeeded_table.get_entry_attrs(succeeded_entries[TEST_CONST.TEST_DATA_FILE_2_NAME][0])
-            for attr_name in ['filename','rel_filepath','n_chunks','first_message'] :
-                self.assertEqual(succeeded_entry_attrs[attr_name],testing_entry_attrs[attr_name])
         except Exception as e :
             raise e
         finally :
