### 2022-10-27 4ba4252
expose logger levels as command line options
and several adjustments to the logging messages themselves
