+++ openmsistream/metadata_extraction/metadata_json_reproducer.py
@@ -100,6 +100,7 @@ class MetadataJSONReproducer(DataFileStreamReproducer,ABC) :
                                           n_producer_threads=args.n_producer_threads,
                                           output_dir=args.output_dir,
                                           update_secs=args.update_seconds,
+                                          streamlevel=args.logger_stream_level,filelevel=args.logger_file_level,
                                           )
         # cls.bucket_name = args.bucket_name
         msg = f'Listening to the {args.consumer_topic_name} topic for XRD CSV files to send their metadata to the '
@@ -107,13 +108,21 @@ class MetadataJSONReproducer(DataFileStreamReproducer,ABC) :
         xrd_csv_metadata_reproducer.logger.info(msg)
         n_r,n_p,f_r_fns,m_p_fns = xrd_csv_metadata_reproducer.produce_processing_results_for_files_as_read()
         xrd_csv_metadata_reproducer.close()
-        msg = f'{n_r} total message{"s were" if n_r!=1 else " was"} consumed'
-        msg+=f', {n_p} message{"s were" if n_p!=1 else " was"} successfully processed'
-        msg+=f', {len(f_r_fns)} file{"s were" if len(f_r_fns)!=1 else " was"} fully-read'
+        msg = ''
+        if n_r>0 :
+            msg+= f'{n_r} total message{"s were" if n_r!=1 else " was"} consumed, '
+        if n_p>0 :
+            msg+=f'{n_p} message{"s were" if n_p!=1 else " was"} successfully processed, '
+        if len(f_r_fns)>0 :
+            msg+=f'{len(f_r_fns)} file{"s were" if len(f_r_fns)!=1 else " was"} fully read, '
         if len(m_p_fns)>0 :
-            msg+=f', and the following {len(m_p_fns)} file'
+            msg+=f'{len(m_p_fns)} file{"s were" if len(m_p_fns)!=1 else " was"} files had json metadata produced'
+            msg+=f' to the "{args.producer_topic_name}" topic, '
+        xrd_csv_metadata_reproducer.logger.info(msg[:-2])
+        if len(m_p_fns)>0 :
+            msg=f'the following {len(m_p_fns)} file'
             msg+=' had' if len(m_p_fns)==1 else 's had'
             msg+=f' json metadata produced to the {args.producer_topic_name} topic:'
             for fn in m_p_fns :
                 msg+=f'\n\t{fn}'
-        xrd_csv_metadata_reproducer.logger.info(msg)
+            xrd_csv_metadata_reproducer.logger.debug(msg)
