+++ test/unittests/config.py
@@ -1,6 +1,6 @@
 #imports
 import os, pathlib
-from openmsistream.workflow.config import RUN_CONST
+from openmsistream.utilities.config import RUN_CONST
 from openmsistream.data_file_io.config import RUN_OPT_CONST
 
 class TestRoutineConstants :

[…]
