### 2022-10-27 b7ee254
only store the relative filepath in the producer dataclass table logs
