### 2022-10-27 f20a996
bugfixes for CI tests
