### 2022-10-27 39785c4
fix a bug in converting the logger levels
caused problems installing services
