### 2022-10-27 bdfeee9
add a wait to a CI test
