### 2022-10-28 #4837 @pibion
@lucask07
@briatte
@ixjlyons
@sptennak
@SarthakJariwala

👋 would you be interested in reviewing this submission for JOSS?  The goal of review is to make sure that the software and paper meet a minimum standard of best practice.  Reviewers go through the checklist at https://joss.readthedocs.io/en/latest/review_checklist.html and work with the author to improve anything that doesn't already meet the requirements.
