### 2022-10-31 #4837 @danielskatz
@pibion - these questions should first go to the track editor (which is me in this case, but is @openjournals/csism-eics more generally for this track).  In this case, there is no COI since the collaboration was outside of the COI time period
