### 2022-11-08 #4896 @lucask07
Installation for a MAC with Intel chip was checked and proceeded as outlined in the [documentation](https://openmsistream.readthedocs.io/en/latest/installation/mac_intel.html).
