### 2022-11-08 #4896 @lucask07
@eminizer
Could you point me to the simplest example to try this package on my own system(s)?
