### 2022-11-08 #4896 @eminizer
Hi @lucask07,

Yes! Certainly the simplest program would be "[UploadDataFile](https://openmsistream.readthedocs.io/en/latest/user_info/main_programs/upload_data_file.html)", which will break a given file into chunks and Produce those chunks as messages to a Kafka topic.

To run that program (or any others), though, you will need access to a Kafka cluster as enumerated in the "[external requirements](https://openmsistream.readthedocs.io/en/latest/installation/getting_started.html#external-requirements)" section of the documentation. It's fairly quick and easy to set up a Kafka cluster to use on [Confluent Cloud](https://confluent.cloud/), though there is a subscription fee to use that service.

If you'd like, for purposes of the review, maybe I could send you the access keys to our group's cluster and you could try passing data through there? Using our cluster would also make it very easy for you to run [our CI tests](https://openmsistream.readthedocs.io/en/latest/dev_info/ci_testing.html) on your own machine once you set the server/username/password as environment variables on your system(s).

Sorry for the trouble; this package is designed to simplify passing streaming data through Kafka, but users generally need to have their own external access to a broker to use. Let me know if you'd like me to send you our access keys through something more secure.

Thanks for your efforts,
Maggie
