### 2022-11-09 #4896 @lucask07
Thanks @eminizer I would like to try an actual test. This would help my understanding of the overall architecture. If there is an _easy_ way for you to allow me to access your cluster I would try passing data through there. However, if this becomes a project in itself let's not bother.
My email is koerner   dot  lucas @ stthomas dot edu.

Thank you, Lucas
