### 2022-11-09 #4896 @lucask07
A paper comment. An example of a real dataset would be helpful. What are typical file types and sizes? Is there a typical rate at which the data is generated/consumed?

And a related side question: is there a guideline for maximum filesize when using Apache Kafka?
