### 2022-11-09 #4896 @eminizer
Thank you, @lucask07, and again sorry for the trouble. Some responses in line:

> I would like to try an actual test. This would help my understanding of the overall architecture. If there is an easy way for you to allow me to access your cluster I would try passing data through there.

Of course! We would definitely want you to try running the code yourself, that's why we're here : )

I wasn't sure how the review would proceed, but using our cluster would definitely be fastest and easiest. I've just emailed you the information for how to connect to it.

There are a number of topics there for the CI tests (listed [here](https://github.com/openmsi/openmsistream/blob/main/test/unittests/config.py#L13-L20),  and the default topic called "test" also exists), but I just created another one called "joss_review" for you to use however you'd like. I set the retention time on that new topic to an hour, so it'll clear itself out regularly, and I otherwise left all the settings at their defaults (6 partitions, ~1MB max message size, etc.). If you would like me to create any other topics for you to use just let me know.

> An example of a real dataset would be helpful. What are typical file types and sizes?

Good questions. We've deployed OpenMSIStream in a variety of lab environments, so some examples of the file data we're moving include .csv and .raw data files, bitmap and .tif images, .xml and .txt description files, and even some proprietary file formats for automated backups. These files range in size from about 1.5KB to about 4.5GB. We also move data that don't come originally from files, in the form of JSON-formatted strings of extracted metadata information.

The PARADIM group with whom we work hosts a number of public domain datasets at [https://data.paradim.org/](https://data.paradim.org/) if you would like to use any of those as examples of real-world data. Other than that, though, the software is generally agnostic to the type (and, to a point, size, see below) of files, so you should be able to test with pretty much anything.

> Is there a typical rate at which the data is generated/consumed?

Generally upload and download speeds are limited mostly by the network and the read/write speed of the disks on which the filess are stored, so we typically get upload and download rates of 60MB/s or so on wide average. We usually set up a [DataFileUploadDirectory](https://openmsistream.readthedocs.io/en/latest/user_info/main_programs/data_file_upload_directory.html) on different lab machines to passively wait for new files to be added, so the exact rate of data production/consumption depends on when experiments are performed and new data are saved.

> is there a guideline for maximum filesize when using Apache Kafka?

There is (theoretically) no limit to the size of _files_, because OpenMSIStream breaks files of any size into individual chunks, and each chunk is Produced as a message. Kafka limits the sizes of individual messages to 1MB by default, and [OpenMSIStream creates messages of about 16KB each by default](https://github.com/openmsi/openmsistream/blob/main/openmsistream/data_file_io/config.py#L32), but these are both configurable through Kafka and OpenMSIStream, respectively. When we're moving very large files we use our own on-prem broker where we can set the max message size on the relevant topic to 32MB, and then we configure OpenMSIStream to send messages of 8-17MB each.

In practice, however, there is more latency through both Kafka and OpenMSIStream when using larger message sizes, and reconstructed files are held in memory unless they're being written to disk, so those are both considerations. I think those 4.5GB files I mentioned are starting to push the edge of what is practically feasible on standard hardware.

Let me know if there's any other way I can be helpful, and thank you!
Maggie
