### 2022-11-14 e474ebc
fix a bug in restarting DataFileUploadDirectories
