### 2022-11-15 8b7cbb2
use a more sensible default chunk size across the board
and update some formatting stuff
