+++ test/unittests/config.py
@@ -7,9 +7,9 @@ class TestRoutineConstants :
     """
     constants used in running tests
     """
-    @property
-    def TEST_TOPIC_NAMES(self) :
-        return {
+    
+    #names of topics on the test cluster to use for each test
+    TEST_TOPIC_NAMES = {
             'test_data_file_directories':'test_data_file_directories',
             'test_data_file_directories_encrypted':'test_oms_encrypted',
             'test_data_file_stream_processor_kafka':'test_data_file_stream_processor',
@@ -19,119 +19,58 @@ class TestRoutineConstants :
             'test_serialization':'test_oms_encrypted',
             'test_metadata_reproducer':'test_metadata_extractor',
         }
-    @property
-    def TEST_ENDPOINT_URL(self) :  # the endpoint_url for s3 bucket connection
-        return os.environ['ENDPOINT_URL']
-    @property
-    def TEST_BUCKET_NAME(self) :  # the bucket name to upload data to (S3Client)
-        return os.environ['BUCKET_NAME']
-    @property
-    def TEST_ASSCESS_KEY_ID(self) :  # the access_key_id for s3 Authentication
-        return os.environ['ACCESS_KEY_ID']
-    @property
-    def TEST_SECRET_KEY_ID(self) :  # the secret_key_id for s3 Authentication
-        return os.environ['SECRET_KEY_ID']
-    @property
-    def TEST_REGION(self) :  # the region for the testing s3 bucket
-        return os.environ['REGION']
-    @property
-    def TEST_CONFIG_FILE_PATH(self) : # The path to the Kafka config file to use
-        return (RUN_CONST.CONFIG_FILE_DIR / f'{RUN_OPT_CONST.DEFAULT_CONFIG_FILE}{RUN_CONST.CONFIG_FILE_EXT}').resolve()
-    @property
-    def TEST_CONFIG_FILE_PATH_ENCRYPTED(self) : # Same as above except it includes a node_id to test encryption
-        return (RUN_CONST.CONFIG_FILE_DIR / f'test_encrypted{RUN_CONST.CONFIG_FILE_EXT}').resolve()
-    @property
-    def TEST_CONFIG_FILE_PATH_ENCRYPTED_2(self) : # Same as above except it includes a node_id to test encryption
-        return (RUN_CONST.CONFIG_FILE_DIR / f'test_encrypted_2{RUN_CONST.CONFIG_FILE_EXT}').resolve()
-    @property
-    def TEST_CONFIG_FILE_PATH_S3_TRANSFER(self) : # Same as above except it includes S3 transfer configs
-        return (RUN_CONST.CONFIG_FILE_DIR / f'test_s3_transfer{RUN_CONST.CONFIG_FILE_EXT}').resolve()
-    @property
-    def PROD_CONFIG_FILE_PATH(self) : # The path to the "prod" Kafka config file to use in making sure that the prod environment variables are not set
-        return (RUN_CONST.CONFIG_FILE_DIR / f'prod{RUN_CONST.CONFIG_FILE_EXT}').resolve()
-    @property
-    def TEST_DIR_PATH(self) :
-        return pathlib.Path(__file__).parent.parent
-    @property
-    def PACKAGE_ROOT_DIR(self) :
-        return self.TEST_DIR_PATH.parent / 'openmsistream'
-    @property
[…]
