+++ test/unittests/test_s3_transfer_stream_processor.py
@@ -31,7 +31,7 @@ class TestS3TransferStreamProcessor(unittest.TestCase):
         upload_thread = ExceptionTrackingThread(target=dfud.upload_files_as_added,
                                                 args=(TOPIC_NAME,),
                                                 kwargs={'n_threads': RUN_OPT_CONST.N_DEFAULT_UPLOAD_THREADS,
-                                                        'chunk_size': RUN_OPT_CONST.DEFAULT_CHUNK_SIZE,
+                                                        'chunk_size': TEST_CONST.TEST_CHUNK_SIZE,
                                                         'max_queue_size': RUN_OPT_CONST.DEFAULT_MAX_UPLOAD_QUEUE_SIZE,
                                                         'upload_existing': True})
         upload_thread.start()
@@ -194,7 +194,7 @@ class TestS3TransferStreamProcessor(unittest.TestCase):
         endpoint_url = TEST_CONST.TEST_ENDPOINT_URL
         if not endpoint_url.startswith('https://'):
             endpoint_url = 'https://' + endpoint_url
-        aws_access_key_id = TEST_CONST.TEST_ASSCESS_KEY_ID
+        aws_access_key_id = TEST_CONST.TEST_ACCESS_KEY_ID
         aws_secret_access_key = TEST_CONST.TEST_SECRET_KEY_ID
         region_name = TEST_CONST.TEST_REGION
         bucket_name = TEST_CONST.TEST_BUCKET_NAME
