+++ test/unittests/test_data_file_stream_processor_encrypted.py
@@ -38,7 +38,7 @@ class TestDataFileStreamProcessorEncyrpted(unittest.TestCase) :
         upload_thread = ExceptionTrackingThread(target=dfud.upload_files_as_added,
                                                 args=(TOPIC_NAME,),
                                                 kwargs={'n_threads':RUN_OPT_CONST.N_DEFAULT_UPLOAD_THREADS,
-                                                        'chunk_size':16*RUN_OPT_CONST.DEFAULT_CHUNK_SIZE,
+                                                        'chunk_size':16*TEST_CONST.TEST_CHUNK_SIZE,
                                                         'max_queue_size':RUN_OPT_CONST.DEFAULT_MAX_UPLOAD_QUEUE_SIZE,
                                                         'upload_existing':True}
                                 )
