### 2022-11-16 86a677a
wait for "Network Connections" service to be available before restarting Windows Services
