### 2022-11-16 bce5400
solve a problem where BufferErrors would still be logged as Errors even though they're expected
also a couple more cosmetics in logging messages
