+++ openmsistream/utilities/controlled_process_multi_threaded.py
@@ -104,9 +104,8 @@ class ControlledProcessMultiThreaded(ControlledProcess,ABC) :
             if thread.caught_exception is not None :
                 #log the error
                 warnmsg = 'WARNING: a thread raised an Exception, which will be logged as an error below but not '
-                warnmsg+= 'reraised. The thread that raised the error will be restarted.'
-                self.logger.warning(warnmsg)
-                self.logger.log_exception_as_error(thread.caught_exception,reraise=False)
+                warnmsg+= 're-raised. The thread that raised the error will be restarted.'
+                self.logger.warning(warnmsg,exc_info=thread.caught_exception)
                 #try to join the thread
                 try :
                     thread.join()
