### 2022-11-16 76075bf
update defaults for the internal upload queue size and for the default batch sizes
