### 2022-11-16 8f3bfbd
dynamically adjust number of items in internal producer queue to keep a static size/usage of RAM
also bump down the batch sizes for encrypted tests
