+++ openmsistream/data_file_io/config.py
@@ -30,6 +30,6 @@ class RunOptionConstants :
     DEFAULT_UPLOAD_REGEX = re.compile(r'^((?!(\.|.*.log))).*$')
 
     DEFAULT_CHUNK_SIZE = 524288 # default size in bytes of each file upload chunk
-    DEFAULT_MAX_UPLOAD_QUEUE_SIZE = 1000 # default maximum number of items allowed in the upload Queue at once
+    DEFAULT_MAX_UPLOAD_QUEUE_MEGABYTES = 500 # default maximum size (in MB) of the internal upload Queue
 
 RUN_OPT_CONST = RunOptionConstants()
