### 2022-11-16 5527b96
fix pattern for logging errors through the custom logger
now much more intuitive by interpreting the "exc_info" keyword argument rather than having a separate argument
