+++ docs/source/user_info/main_programs/data_file_upload_directory.rst
@@ -24,7 +24,7 @@ Other options for running the code include:
 #. Changing which files in the directory get uploaded: add the ``--upload_regex [regex]`` argument where ``[regex]`` is a regular expression. Only files whose names match ``[regex]`` will be uploaded. The default excludes any hidden files (names start with '.') and log files (names end with '.log').
 #. Uploading existing files as well as newly-added files: add the ``--upload_existing`` flag (see above).
 #. Changing the size of the individual file chunks: add the ``--chunk_size [n_bytes]`` argument where ``[n_bytes]`` is the desired chunk size in bytes. ``[n_bytes]`` must be a nonzero power of two (the default is 16384). It may be necessary to reconfigure the topic you're using as well as some of the Producer configurations to allow particularly large messages.
-#. Changing the number of messages that are allowed to be internally queued at once (that is, queued before being produced): add the ``--queue_max_size [n_messages]`` argument where ``[n_messages]`` is the desired number of messages allowed in the internal queue (the default is 3000 messages). This internal queue is used to make sure that there's some buffer between recognizing a file exists to be uploaded and producing all of its associated messages to the topic; its size should be set to some number of messages such that the total size of the internal queue is capped at a few batches of messages ("``batch.size``" in the producer config). The default value is reasonable.
+#. Changing the size of the internal message queue: add the ``--queue_max_size [n_megabytes]`` argument where ``[n_megabytes]`` is the maximum allowed size of the internal queue in megabytes (the default is 500MB). This internal queue is used to make sure that there's some buffer between recognizing a file exists to be uploaded and producing all of its associated messages to the topic; its size should be set somewhere between a few batches of messages ("``batch.size``" in the producer config) and the maximum memory footprint you would like the program to have. The default value is 500MB. If possible, it's preferable to set this value at least as large as the librdkafka buffering queue ("``queue.buffering.max.kbytes``" in the producer config) to minimize latency, but the total memory footprint is approximately the sum of both queues, and the default librdkafka queue size is about 1GB.
 
 To see other optional command line arguments, run ``DataFileUploadDirectory -h``.
 
