### 2022-11-21 0f1701a
fixed a bug preventing module loading if env vars not set
