### 2022-11-21 #4896 @lucask07
Hello @eminizer

I have added the Kafka environmental variables and now am moving further with the tests. Progress!

```
Ran 49 tests in 175.359s

FAILED (failures=2, errors=4, skipped=4)
```
Seems that most of the issues are FileExistsError:

`FileExistsError: [Errno 17] File exists: '/somepaths/joss/openMSIstream/openmsistream/test/test_reco'
`
After removing these files the tests advanced further

```
test_initial_properties (test_upload_data_file.TestUploadDataFile) ... ok
test_upload_whole_file_kafka (test_upload_data_file.TestUploadDataFile) ... ok

FAIL: test_data_file_stream_processor_restart_kafka (test_data_file_stream_processor.TestDataFileStreamProcessor)
Test restarting a DataFileStreamProcessor from the beginning of the topic after failing to process a file
Traceback (most recent call last):
  File "/somepaths/joss/openMSIstream/openmsistream/test/unittests/test_data_file_stream_processor.py", line 132, in test_data_file_stream_processor_restart_kafka
    self.assertFalse(TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR_RESTART.is_dir())
AssertionError: True is not false

FAIL: test_data_file_stream_processor_restart_encrypted_kafka (test_data_file_stream_processor_encrypted.TestDataFileStreamProcessorEncyrpted)
Test restarting an encrypted DataFileStreamProcessor after failing to process a file
Traceback (most recent call last):
  File "/somepaths/joss/openMSIstream/openmsistream/test/unittests/test_data_file_stream_processor_encrypted.py", line 26, in test_data_file_stream_processor_restart_encrypted_kafka
    self.assertFalse(TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR_RESTART_ENCRYPTED.is_dir())
AssertionError: True is not false

Ran 49 tests in 200.716s

FAILED (failures=2, skipped=4)
```
