### 2022-11-21 #4896 @lucask07
My tests (on a MAC) were initially throwing an SSL error like:
`[SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed`

I removed this error by adding
```
CERT_PATH=$(python -m certifi)
export SSL_CERT_FILE=${CERT_PATH}
export REQUESTS_CA_BUNDLE=${CERT_PATH}
```
to my bash_profile as described at this [StackOverflow question](https://stackoverflow.com/questions/40684543/how-to-make-python-use-ca-certificates-from-mac-os-truststore).
