### 2022-11-21 openmsi/openmsistream#26 @lucask07
Documentation: configuration files
The documentation states:

> Working with any of the programs in OpenMSIStream requires creating at least one configuration file that will tell the Kafka backend how to connect to the broker and configure the main Producer(s)/Consumer(s) it uses. Information in configuration files is also used to point a Consumer to an S3 bucket to the correct endpoint and connect to it, and to supply information necessary for working with data encrypted using KafkaCrypto.

Where do these configuration files go? How do I decide which one OpenMISStream is using?
