### 2022-11-21 openmsi/openmsistream#27 @lucask07
Documentation: simplest possible example
Could the repository include an 'examples' folder that contains some very simple uses? Possibly this includes both a "producer" and a "consumer"?
