### 2022-11-21 #4896 @lucask07
The paper reads very well. Nice work!

A few boxes I will wait for feedback to check off:

- The package would benefit from a few practical examples added to the docs and repository. These examples ideally would be a tangible experiment.  [GitHub Issue #27](https://github.com/openmsi/openmsistream/issues/27)
- It is not clear how to use the configuration files. [GitHub Issue #26](https://github.com/openmsi/openmsistream/issues/26)

The
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

in the paper could use a revisit. The authors mention and cite Bluesky from NSLS but do not note if this fulfills the need. Are there other Python packages that enable streaming of data?

Hopefully, I can get all the tests working as well!
