### 2022-11-21 #4896 @eminizer
Hi @lucask07,

Thanks for the detailed look!

First, in terms of the tests, I just released a new version on PyPi (v1.1.6) that should prevent the failures you were seeing. If you update your install and try again hopefully you'll have more success this time!

Second, I love your idea of including concrete examples in the repo and documentation. I'll absolutely work on this soon, and I'll link a PR to the issue you opened for it. I'll also do the same for some documentation updates to better explain the usage of the configuration files. These later updates might take a bit longer due to the short week, but I'll get to them as soon as I'm able.

We can also flesh out that section in the paper to provide more context.

Thanks again for all of your attention to detail and for your help in improving our body of work : )

Best,
Maggie
