### 2022-11-22 openmsi/openmsistream#26 @eminizer
The new section I added to clarify where config files can go and how users can set which one is used is called "Location and Usage" and can be found [here](https://openmsistream.readthedocs.io/en/latest/user_info/main_programs.html#location-and-usage) on readthedocs : )
