### 2022-11-22 13a0ce2
add an entry in the troubleshooting docs about the SSL certificate verification error
