+++ docs/source/introduction/tutorials.rst
@@ -1,0 +1,65 @@
+=========
+Tutorials
+=========
+
+The hands-on examples described in this section will walk through some of the key functionality of the OpenMSIStream package. After setting up access to a Kafka cluster, creating a topic to use for testing, and downloading some specific example data, you'll use a :doc:`DataFileUploadDirectory <../user_info/main_programs/data_file_upload_directory>` producer to upload the example files from your local machine to the topic. You'll then use a :doc:`DataFileDownloadDirectory <../user_info/main_programs/>` consumer to reconstruct the files to your local machine from their chunks stored on the topic.
+
+More advanced tutorials explore other OpenMSIStream functionality by setting up additional types of consumers on your local machine. An example :doc:`DataFileStreamProcessor <../user_info/base_classes/data_file_stream_processor>` can create simple plots from the data in the files on the topic.  An example :doc:`MetadataJSONReproducer <../user_info/base_classes/metadata_json_reproducer>` can extract the metadata blocks of the example files and produce those blocks to a second topic as JSON-formatted strings. An :doc:`S3TransferStreamProcessor <../user_info/main_programs/s3_transfer_stream_processor>` can reconstruct the files in an S3 bucket instead of locally on disk (after some more setup).
+
+Initial setup
+=============
+
+To get started, you'll need to get access to a Kafka broker, create at least two topics to use for testing, set some related environment variables (in order to use the example config files), and download some data to use in your tests.
+
+Access to a Kafka broker
+------------------------
+
+To begin, you'll need access to a Kafka broker. That broker can be running on a server that you have access to, or on `Confluent Cloud <https://confluent.cloud/>`_, which provides Kafka as a managed service. In either case you will need to be able to authenticate to that broker using plain SASL authentication (the default for Confluent Cloud). 
+
+You will also need to create a topic on that broker to use for testing, called "``openmsi_tutorial_data``" (if you use a different name just replace it in the commands shown below). The default settings are fine for a single test, but if you'd like you can adjust the retention times to be an hour or less so that the contents of the testing topics will get flushed out regularly and you can work through the tutorial repeatedly without reading previously-produced data.
+
+Setting environment variables
+-----------------------------
+
+Next you'll need to set some environment variables needed to connect to the Kafka broker. You'll need the "Bootstrap server" for the cluster, as well as an API key username and password. If you're using Confluent Cloud, you can find the bootstrap server on the "Cluster settings" page from the left hand menu, and you can create a new API key using the "API keys" page also on the left hand menu. Set the corresponding values for the "``KAFKA_TEST_CLUSTER_BOOTSTRAP_SERVERS``", "``KAFKA_TEST_CLUSTER_USERNAME``" and "``KAFKA_TEST_CLUSTER_PASSWORD``" environment variables on your system [#]_.
+
+Downloading test data
+---------------------
+
+Lastly, you'll need to download some test data to work with. The five files linked below contain X-Ray Diffraction (XRD) data made publicly available under the `CC BY-NC-ND 4.0 <https://creativecommons.org/licenses/by-nc-nd/4.0/>`_ license by the `PARADIM Materials Innovation Platform <https://www.paradim.org/>`_, as part of the work leading to the publication of S. Chae et. al., "Epitaxial stabilization of rutile germanium oxide thin film by molecular beam epitaxy", Applied Physics Letters 117, 072105 (2020) `https://doi.org/10.1063/5.0018031 <https://doi.org/10.1063/5.0018031>`_. You can download these five files to your machine to use them in this tutorial.
+
[…]
