+++ docs/source/introduction/tutorials/extracting_metadata.rst
@@ -1,0 +1,3 @@
+====================================================
+Extracting metadata (example MetadataJSONReproducer)
+====================================================
