+++ docs/source/introduction/tutorials/creating_plots.rst
@@ -1,0 +1,3 @@
+================================================
+Creating plots (example DataFileStreamProcessor)
+================================================
