### 2022-11-30 e9a3c17
adding documentation for the first tutorial
and some updates to make the outputs a bit more readable
