+++ test/unittests/test_data_file_directories_encrypted.py
@@ -76,7 +76,8 @@ class TestDataFileDirectories(unittest.TestCase) :
             LOGGER.info(msg)
             reco_fp = TEST_CONST.TEST_RECO_DIR_PATH_ENCRYPTED/TEST_CONST.TEST_DATA_FILE_SUB_DIR_NAME
             reco_fp = reco_fp/TEST_CONST.TEST_DATA_FILE_NAME
-            while (reco_fp not in dfdd.completely_processed_filepaths) and time_waited<TIMEOUT_SECS :
+            reco_rel_fp = reco_fp.relative_to(TEST_CONST.TEST_RECO_DIR_PATH_ENCRYPTED)
+            while (reco_rel_fp not in dfdd.completely_processed_filepaths) and time_waited<TIMEOUT_SECS :
                 current_messages_read = dfdd.n_msgs_read
                 LOGGER.info(f'\t{current_messages_read} messages read after waiting {time_waited} seconds....')
                 time.sleep(5)
