### 2022-11-30 32060d1
bugfixes for the relative paths actually being paths and not strings now
