### 2022-11-30 076c3d3
move around the metadata extraction example
and get rid of the test file in the repository
