+++ docs/source/user_info/base_classes/metadata_json_reproducer.rst
@@ -8,8 +8,6 @@ MetadataJSONReproducer
 
 A MetadataJSONReproducer is a base class that can be extended to extract a dictionary of metadata keys and values from data files as they're reconstructed from a Kafka topic, and then "re-produce" those metadata dictionaries to a different topic as JSON-formatted strings. A user only needs to write one function in classes extending MetadataJSONReproducer to run the workflow defined by this base class. 
 
-An example of an implemented MetadataJSONReproducer-type class can be found `in the "metadata_extraction/examples" subdirectory of the OpenMSIStream repository <https://github.com/openmsi/openmsistream/tree/main/openmsistream/metadata_extraction/examples>`_. That directory also contains an example configuration file to use and a test data .csv file containing metadata in a header block. You can upload the test file to a topic with UploadDataFile, and then run the XRDCSVMetadataReproducer class as a module to reconstruct the test file from the topic you uploaded it to, extract the metadata as a dictionary, and produce it to another topic.
-
 The main function that users will need to write to instantiate their own MetadataJSONReproducer-type classes is :func:`openmsistream.MetadataJSONReproducer._get_metadata_dict_for_file`, which will compute the metadata dictionary for each reconstructed data file as they become available. The argument to that function is an :class:`~.data_file_io.DownloadDataFileToMemory` object holding a file that has been fully reconstructed from chunks in a topic; the content of the file can be accessed using its :attr:`bytestring` attribute. The function should return a dictionary of metadata keys and values when successful, but it can safely raise Exceptions or log warnings if necessary (Exceptions raised will be logged without crashing the program).
 
 If a user would like to add additional command line arguments for the extended class they can also implement the :func:`openmsistream.utilities.Runnable.get_command_line_arguments` class method. Command line arguments already defined include the config file, consumer and producer topic names/numbers of threads, and output location; they can be picked up by calling :func:`super().get_command_line_arguments` in the extended class.
