### 2022-12-01 b9a51a4
add tutorial for automated metadata extraction
