### 2022-12-01 87a7878
add a tutorial for creating some simple scatterplots from the test files
