### 2022-12-01 7a8d725
bugfixes for relocated stream processor logs
