### 2022-12-02 ede0d02
add a tutorial for running an S3TransferStreamProcessor
