### 2022-12-02 306ef86
add some READMEs explaining what's in the "examples" subdirectory
