### 2022-12-02 932ec47
add a CI test for the plotting tutorial
