+++ test/unittests/test_plots_for_tutorial.py
@@ -1,0 +1,134 @@
+#imports
+import unittest, pathlib, logging, time, shutil, urllib.request
+import importlib.machinery, importlib.util
+from openmsistream import UploadDataFile
+from openmsistream.utilities import Logger
+from openmsistream.utilities.exception_tracking_thread import ExceptionTrackingThread
+from openmsistream.data_file_io.actor.file_registry.stream_handler_registries import StreamProcessorRegistry
+from config import TEST_CONST
+
+#import the XRDCSVPlotter from the examples directory
+class_path = TEST_CONST.EXAMPLES_DIR_PATH / 'creating_plots' / 'xrd_csv_plotter.py'
+module_name = class_path.name[:-len('.py')]
+loader = importlib.machinery.SourceFileLoader(module_name, str(class_path))
+module = loader.load_module()
+
+#constants
+LOGGER = Logger(pathlib.Path(__file__).name.split('.')[0],logging.ERROR)
+TIMEOUT_SECS = 90
+JOIN_TIMEOUT_SECS = 60
+UPLOAD_FILE = TEST_CONST.EXAMPLES_DIR_PATH / 'creating_plots' / 'SC001_XRR.csv'
+TOPIC_NAME = TEST_CONST.TEST_TOPIC_NAMES[pathlib.Path(__file__).name[:-len('.py')]]
+CONSUMER_GROUP_ID = 'test_plots_for_tutorial'
+
+class TestPlotsForTutorial(unittest.TestCase) :
+    """
+    Class for testing that an uploaded file can be read back from the topic and have its metadata 
+    successfully extracted and produced to another topic as a string of JSON
+    """
+
+    def setUp(self) :
+        """
+        Download the test data file from its URL on the PARADIM website
+        """
+        urllib.request.urlretrieve(TEST_CONST.TUTORIAL_TEST_FILE_URL,
+                                   UPLOAD_FILE)
+
+    def tearDown(self) :
+        """
+        Remove the test data file that was downloaded
+        """
+        if UPLOAD_FILE.is_file() :
+            UPLOAD_FILE.unlink()
+
+    def test_plots_for_tutorial_kafka(self) :
+        #remove any old output
+        if TEST_CONST.TEST_TUTORIAL_PLOTS_OUTPUT_DIR.is_dir() :
+            shutil.rmtree(TEST_CONST.TEST_TUTORIAL_PLOTS_OUTPUT_DIR)
+        #start up the plot maker
+        plot_maker = module.XRDCSVPlotter(
+                                TEST_CONST.TEST_CONFIG_FILE_PATH,
+                                TOPIC_NAME,
+                                consumer_group_id=CONSUMER_GROUP_ID,
+                                output_dir=TEST_CONST.TEST_TUTORIAL_PLOTS_OUTPUT_DIR,
+                                logger=LOGGER,
+                            )
+        plotting_thread = ExceptionTrackingThread(target=plot_maker.process_files_as_read)
+        plotting_thread.start()
+        try :
+            #upload the test data file
+            LOGGER.set_stream_level(logging.INFO)
+            msg = 'Uploading test plotting file in test_plots_for_tutorial'
+            LOGGER.info(msg)
+            LOGGER.set_stream_level(logging.ERROR)
+            upload_file = UploadDataFile(UPLOAD_FILE,
+                                        logger=LOGGER)
+            upload_file.upload_whole_file(TEST_CONST.TEST_CONFIG_FILE_PATH,TOPIC_NAME)
+            #wait for the file to be processed
+            current_messages_read = -1
+            time_waited = 0
+            LOGGER.set_stream_level(logging.INFO)
+            msg = f'Waiting to create plots from the "{TOPIC_NAME}" topic in '
[…]
