### 2022-12-02 72ce6f1
allow using a custom "services" working directory based on an environment variable
