+++ openmsistream/services/config.py
@@ -1,7 +1,7 @@
 """Utility class holding constants and small calculations for working with Services/daemons"""
 
 #imports
-import pathlib, importlib
+import os, pathlib, importlib
 from inspect import isclass
 import pkg_resources
 from ..utilities import Logger
@@ -11,7 +11,9 @@ class ServicesConstants :
     Constants for working with services
     """
 
-    WORKING_DIR = (pathlib.Path(__file__).parent/'working_dir').resolve()
+    WORKING_DIR = ( os.environ['OPENMSISTREAM_SERVICES_WORKING_DIR'] 
+                    if 'OPENMSISTREAM_SERVICES_WORKING_DIR' in os.environ else 
+                    (pathlib.Path(__file__).parent/'working_dir').resolve() )
     NSSM_PATH = WORKING_DIR / 'nssm.exe'
     NSSM_DOWNLOAD_URL = 'https://nssm.cc/release/nssm-2.24.zip' # The URL to use for downloading NSSM when needed
     ERROR_LOG_STEM = '_ERROR_LOG.txt'
