+++ docs/source/user_info/services.rst
@@ -62,7 +62,7 @@ If something goes wrong while the program is running as a Service/daemon, a file
 Output in the "working directory"
 ---------------------------------
 
-Working with Services will create a few files in the `working_dir subdirectory of the OpenMSIStream repo <https://github.com/openmsi/openmsistream/tree/main/openmsistream/services/working_dir>`_. A logfile called "Services.log" will contain some lines related to installing or working with any Services or daemons. Python files in that directory will correspond to the executables that are installed as Services, and so checking these files will give details on the setup for their corresponding Services/daemons. None of these created files will be tracked in the repo (`they're in the .gitignore <https://github.com/openmsi/openmsistream/blob/main/.gitignore>`_).
+Working with Services will create a few files in a "working directory". The default working directory is the `working_dir subdirectory of the OpenMSIStream repo <https://github.com/openmsi/openmsistream/tree/main/openmsistream/services/working_dir>`_, but you can change its location by setting the ``OPENMSISTREAM_SERVICES_WORKING_DIR`` environment variable on your system. A logfile called "Services.log" will contain some lines related to installing or working with any Services or daemons. Python files in that directory will correspond to the executables that are installed as Services, and so checking these files will give details on the setup for their corresponding Services/daemons. The directory may also contain text files listing environment variables values or installation arguments for services that you install. None of these created files will be tracked in the repo if they're in the default location (`they're in the .gitignore <https://github.com/openmsi/openmsistream/blob/main/.gitignore>`_).
 
 Running other Python code as Services/Daemons
 ---------------------------------------------
