+++ openmsistream/utilities/config.py
@@ -1,7 +1,7 @@
 """Constants used in running different controlled processes in general"""
 
 #imports
-import pathlib
+import os, pathlib
 
 class RunConstants :
     """
@@ -9,7 +9,9 @@ class RunConstants :
     """
 
     CONFIG_FILE_EXT = '.config'
-    CONFIG_FILE_DIR = (pathlib.Path(__file__).parent.parent / 'kafka_wrapper' / 'config_files').resolve()
+    CONFIG_FILE_DIR = ( os.environ['OPENMSISTREAM_CONFIG_FILE_DIR']
+                        if 'OPENMSISTREAM_CONFIG_FILE_DIR' in os.environ else
+                        (pathlib.Path(__file__).parent.parent / 'kafka_wrapper' / 'config_files').resolve() )
     DEFAULT_N_THREADS = 2
     DEFAULT_UPDATE_SECONDS = 300
 
