+++ test/unittests/test_data_file_stream_processor.py
@@ -62,11 +62,11 @@ class TestDataFileStreamProcessor(unittest.TestCase) :
         #upload the data file
         upload_datafile = UploadDataFile(TEST_CONST.TEST_DATA_FILE_2_PATH,
                                          rootdir=TEST_CONST.TEST_DATA_DIR_PATH,logger=LOGGER)
-        upload_datafile.upload_whole_file(TEST_CONST.TEST_CONFIG_FILE_PATH,TOPIC_NAME,
+        upload_datafile.upload_whole_file(TEST_CONST.TEST_CFG_FILE_PATH,TOPIC_NAME,
                                           n_threads=RUN_OPT_CONST.N_DEFAULT_UPLOAD_THREADS,
                                           chunk_size=TEST_CONST.TEST_CHUNK_SIZE)
         #Use a stream processor to read its data back into memory
-        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CONFIG_FILE_PATH,
+        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH,
                                                  TOPIC_NAME,
                                                  output_dir=TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR,
                                                  n_threads=RUN_OPT_CONST.N_DEFAULT_DOWNLOAD_THREADS,
@@ -136,17 +136,17 @@ class TestDataFileStreamProcessor(unittest.TestCase) :
         #upload the data files
         upload_datafile = UploadDataFile(TEST_CONST.TEST_DATA_FILE_PATH,
                                          rootdir=TEST_CONST.TEST_DATA_FILE_ROOT_DIR_PATH,logger=LOGGER)
-        upload_datafile.upload_whole_file(TEST_CONST.TEST_CONFIG_FILE_PATH,TOPIC_NAME,
+        upload_datafile.upload_whole_file(TEST_CONST.TEST_CFG_FILE_PATH,TOPIC_NAME,
                                           n_threads=RUN_OPT_CONST.N_DEFAULT_UPLOAD_THREADS,
                                           chunk_size=TEST_CONST.TEST_CHUNK_SIZE)
         upload_datafile = UploadDataFile(TEST_CONST.TEST_DATA_FILE_2_PATH,
                                          rootdir=TEST_CONST.TEST_DATA_DIR_PATH,logger=LOGGER)
-        upload_datafile.upload_whole_file(TEST_CONST.TEST_CONFIG_FILE_PATH,TOPIC_NAME,
+        upload_datafile.upload_whole_file(TEST_CONST.TEST_CFG_FILE_PATH,TOPIC_NAME,
                                           n_threads=RUN_OPT_CONST.N_DEFAULT_UPLOAD_THREADS,
                                           chunk_size=TEST_CONST.TEST_CHUNK_SIZE)
         #Use a stream processor to read their data back into memory one time, deliberately failing the first file
         time.sleep(1.0)
-        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CONFIG_FILE_PATH,
+        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH,
                                                  TOPIC_NAME,
                                                  output_dir=TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR_RESTART,
                                                  n_threads=RUN_OPT_CONST.N_DEFAULT_DOWNLOAD_THREADS,
@@ -223,12 +223,12 @@ class TestDataFileStreamProcessor(unittest.TestCase) :
         third_filepath = TEST_CONST.FAKE_PROD_CONFIG_FILE_PATH
[…]
