### 2022-12-02 #4896 @eminizer
Hi again @lucask07,

I've just released version 1.2.0 of OpenMSIStream, which includes some extra [documentation](https://openmsistream.readthedocs.io/en/latest/user_info/main_programs.html#location-and-usage) for how to use configuration files see and a brand new [section](https://openmsistream.readthedocs.io/en/latest/introduction/tutorials.html#tutorials) with detailed tutorials for uploading/downloading files, transferring files to S3 buckets, and making plots and extracting metadata from files on streams. The tutorials are covered by CI tests, as well. Please take a look and let me know what you think! If you'd like to try the tutorials, I've created topics with the appropriate names on our Kafka broker and you should be all good to go with then environment variables you've already set.

The edit to the paper text we will likely put together at the start of next week; we'd like to gather a little more context first.

Thanks as always for your continued help and for the improvements to our work that you've inspired : )

Best,
Maggie
