### 2022-12-05 #4896 @lucask07
Hi @eminizer,
The tutorials are very nicely done. Great work! I have checked off the 'Example Usage' box. Unfortunately, with the end of the semester approaching, I won't have time to try these out.

I will check back when the paper edits are posted.

Best,
Lucas
