+++ paper.md
@@ -79,7 +79,9 @@ In materials science projects, it is common to see iterative scientific design w
 
 OpenMSIStream has been adopted as the streaming solution for the Open Material Semantic Infrastructure (OpenMSI) within the NSF-sponsored Designing Materials to Revolutionize and Engineer our Future (DMREF) collaboration at the Hopkins Extreme Materials Institute (HEMI). OpenMSIStream is also currently used to automate data transfer and analysis between electron microscopy laboratories at Cornell University and Johns Hopkins University as part of the NSF-sponsored PARADIM Materials Innovation Platform, and for similar purposes for X-ray instruments at the Materials Characterization and Processing (MCP) Facility at Johns Hopkins University. Further, OpenMSIStream will be deployed in the near future as part of the data streaming solutions for the ARL-sponsored High-Throughput Materials Discovery for Extreme Conditions (HTMDEC) project, as well as in the Integrated Materials Design and Processing Applications to Recycled Plastics DMREF project (NSF).
 
-OpenMSIStream is written in Python and so interfaces seamlessly with other existing components of scientific software stacks such as $\texttt{NumPy}$ [@numpy], $\texttt{SciPy}$ [@scipy], and $\texttt{pandas}$ [@pandas], as well as the Bluesky data collection framework [@bluesky] which also uses Apache Kafka for its streaming backend. The use of the Kafka backend also allows users even more familiar with the Kafka ecosystem to take full advantage of non-Python tools like Kafka Streams [@kafkastreams] for further data handling outside of OpenMSIStream.
+There are many existing open source Python libraries that implement streaming data workflows. For example, Bytewax [@bytewax] is a library for managing data flow with streaming components that can integrate with Kafka, providing an interface to different data sources and generalized operators at the single-message level. The Bluesky data collection framework [@bluesky] implements a DataBroker class that can access local and remote files, registering incremental updates to them as data are collected. Other libraries like Streamz [@streamz] can be used to build components of local streaming architectures that don't access or publish data remotely. 
+
+While OpenMSIStream would integrate well with any of these Python tools, it also fulfills a need at their intersection for something that implements full stream processing capability based around a remote broker while remaining lightweight and accessible to laboratory students and investigators accustomed to writing Python code for analyzing files stored on disk. OpenMSIStream also interfaces seamlessly with other existing components of scientific software stacks such as $\texttt{NumPy}$ [@numpy], $\texttt{SciPy}$ [@scipy], and $\texttt{pandas}$ [@pandas]. The use of the Kafka backend allows users even more familiar with the Kafka ecosystem to take full advantage of non-Python tools like Kafka Streams [@kafkastreams] for further data handling outside of OpenMSIStream.
 
 # Acknowledgments
 
