### 2023-01-19 #4896 @PeterKraus
I am happy to contribute a review.

For a full disclosure, I should mention that we've invited @eminizer to contribute towards our activities over at https://github.com/marda-alliance/metadata_extractors; to me, this is not a conflict of interest, but the editors might see this differently.
