### 2023-02-23 #4896 @eminizer
Yay, thank you everyone!

@SergeyYakubov when you're ready to get started, could you please send me your email (or reach out to me at margaret.eminizer@gmail.com) so that I can send you some secrets you'll need to set as environment variables to run the CI tests and connect to our Kafka cluster on Confluent Cloud?
