### 2023-03-05 openmsi/openmsistream#32 @SergeyYakubov
test issues
Hi, I have collected here several issues related to testing openmsistream

1. lint tests fail for me
`RuntimeError: ERROR: pylint checks failed with output:
************* Module openmsistream
openmsistream/__init__.py:47:16: R1713: Consider using str.join(sequence) for concatenating strings from an iterable (consider-using-join)
************* Module openmsistream.data_file_io.actor.data_file_upload_directory
openmsistream/data_file_io/actor/data_file_upload_directory.py:252:16: R1713: Consider using str.join(sequence) for concatenating strings from an iterable (consider-using-join)
************* Module openmsistream.kafka_wrapper.consumer_group
openmsistream/kafka_wrapper/consumer_group.py:5:0: C0411: third party import "from confluent_kafka.admin import AdminClient" should be placed before "import kafka" (wrong-import-order)
openmsistream/kafka_wrapper/consumer_group.py:6:0: C0411: third party import "from confluent_kafka import TopicPartition" should be placed before "import kafka" (wrong-import-order)
`
2. Flake tests fail for me
`./test.py:2:1: 'openmsistream.data_file_io.config.RUN_OPT_CONST' imported but unused
./test.py:3:1: 'openmsistream.kafka_wrapper.openmsistream_producer.OpenMSIStreamProducer' imported but unused
./test.py:4:1: 'openmsistream.kafka_wrapper.openmsistream_consumer.OpenMSIStreamConsumer' imported but unused
./test.py:5:1: 'openmsistream.kafka_wrapper.producer_group.ProducerGroup' imported but unused
./test.py:7:1: 'unittest' imported but unused
`
3.  `--no_kafka` tests still wants to test connections to Kafka, so they fail if environment variables not set or no service is running
4. Not so important, but it feels wrong to me - you name your tests unit tests, but the ones that require Kafka are actually integration tests. Unit tests should work in isolation without any 3rd party dependencies
