### 2023-03-05 openmsi/openmsistream#33 @SergeyYakubov
Use local Kafka broker
In openmsistream [documentation](https://openmsistream.readthedocs.io/en/latest/introduction/installing_openmsistream.html) and in testing [documentation](https://openmsistream.readthedocs.io/en/latest/dev_info/ci_testing.html) you suggest using commercial service Confluent Cloud for Kafka broker. Actually, it is quite easy to deploy a local Kafka service using Docker compose. With that, a new user can get acknowledged with openmsistream and run tutorials without extra accounts. You could provide a script that helps to deploy Kafka and create necessary topics. A broker could run in no TLS mode, so no certificates/username/password is needed.

Furthermore, local Kafka could be used for local tests as well. Some changes would need to be made to allow testing without TLS.
