### 2023-03-05 #4896 @SergeyYakubov
Hi @eminizer .

great work, openmsistream can be quite useful for a broader community. The paper reads good, no issues there.

I have a couple of issues related to testing: https://github.com/openmsi/openmsistream/issues/32

And a general suggestion to use local Kafka broker (at least document how to do that and maybe provide a couple of scripts): https://github.com/openmsi/openmsistream/issues/33
