### 2023-03-06 #4896 @eminizer
Hi @SergeyYakubov,

Thank you so much for your detailed evaluation and your excellent suggestions. I love your ideas, and I'll get to work on them right now : )

I'll merge some PRs and let you know when we have a next version ready.

Thank you again! And best wishes,
Maggie
