### 2023-03-06 41aec84
fix --no_kafka tests to work even if environment variable values aren't set
Just uses placeholder values
