### 2023-03-06 6934411
temporarily un-set environment variables in question when running test script with "--no_kafka"
