+++ setup.py
@@ -49,6 +49,7 @@ setupkwargs = dict(
                                'marko[toc]',
                                'pyflakes>=2.2.0',
                                'pylint>=2.15.2',
+                               'tempenv>=2.0.0',
                                ],
                       'docs': ['sphinx>=5.0.2',
                                'sphinx_rtd_theme>=1.0.0',
