### 2023-03-06 ec5b1ed
add "no_kafka" tests to the CircleCI suite
should be a more robust check for systems without environment variables set
