### 2023-03-06 b925890
update some dependencies in setup
