+++ test/run_all_tests.py
@@ -2,11 +2,11 @@
 import unittest, subprocess, pathlib, re
 from argparse import ArgumentParser
 from tempenv import TemporaryEnvironment
-from unittests.placeholder_env_vars import ENV_VAR_NAMES
+from test_scripts.placeholder_env_vars import ENV_VAR_NAMES
 
 #constants
 TOP_DIR_PATH = (pathlib.Path(__file__).parent.parent).resolve()
-UNITTEST_DIR_PATH = (pathlib.Path(__file__).parent / 'unittests').resolve()
+TEST_SCRIPT_DIR_PATH = (pathlib.Path(__file__).parent / 'test_scripts').resolve()
 TEST_REPO_STATUS_SCRIPT_PATH = (pathlib.Path(__file__).parent / 'test_repo_status.sh').resolve()
 CWD = pathlib.Path().resolve()
 
@@ -17,13 +17,13 @@ def main(args=None) :
                         help='Add this flag to skip running the pyflakes check')
     parser.add_argument('--no_pylint', action='store_true',
                         help='Add this flag to skip running the pylint checks')
-    unittest_opts = parser.add_mutually_exclusive_group()
-    unittest_opts.add_argument('--no_unittests', action='store_true',
-                               help='Add this flag to skip running the unittest checks')
-    unittest_opts.add_argument('--no_kafka', action='store_true',
-                               help='Add this flag to skip running the unittest checks')
-    unittest_opts.add_argument('--unittest_regex',type=re.compile,default=None,
-                               help='Only unittests whose function names match this regex will be run')
+    script_test_opts = parser.add_mutually_exclusive_group()
+    script_test_opts.add_argument('--no_script_tests', action='store_true',
+                                  help=f'Add this flag to skip running the tests in {TEST_SCRIPT_DIR_PATH.name}')
+    script_test_opts.add_argument('--no_kafka', action='store_true',
+                                  help=f'Add this flag to skip running the tests in {TEST_SCRIPT_DIR_PATH.name}')
+    script_test_opts.add_argument('--test_regex',type=re.compile,default=None,
+                                  help='Only tests whose function names match this regex will be run')
     parser.add_argument('--no_repo', action='store_true',
                         help='Add this flag to skip running the Git repository checks')
     parser.add_argument('--failfast', action='store_true',
@@ -51,13 +51,13 @@ def main(args=None) :
         if stdout!='' :
             raise RuntimeError(f'ERROR: pylint checks failed with output:\n{stdout}')
         print('Passed pylint checks : )')
-    #perform all the unittests
-    if args.no_unittests :
-        print('SKIPPING UNITTESTS')
+    #perform all the script tests
+    if args.no_script_tests :
+        print('SKIPPING SCRIPT TESTS')
     else :
-        print(f'Running unittests in {UNITTEST_DIR_PATH}...')
+        print(f'Running tests in {TEST_SCRIPT_DIR_PATH}...')
         loader = unittest.TestLoader()
-        suites = loader.discover(UNITTEST_DIR_PATH)
+        suites = loader.discover(TEST_SCRIPT_DIR_PATH)
         if len(loader.errors)>0 :
             print('ERROR: encountered the following errors in loading tests:')
             for error in loader.errors :
@@ -79,13 +79,13 @@ def main(args=None) :
                             msg = 'tests that interact with the kafka broker are being skipped'
                             setattr(test, test_name, 
                                     unittest.skip(msg)(getattr(test, test_name)))
-        elif args.unittest_regex is not None :
+        elif args.test_regex is not None :
             for suite in suites :
                 for test_group in suite._tests :
                     for test in test_group :
-                        if not args.unittest_regex.match(test._testMethodName) :
[…]
