### 2023-03-06 fa9f916
add config files, scripts, etc. for a local kafka broker option in running tests
