+++ test/test_scripts/test_data_file_stream_processor_encrypted.py
@@ -37,7 +37,7 @@ class TestDataFileStreamProcessorEncyrpted(unittest.TestCase) :
         watched_subdir.mkdir(parents=True)
         #start up the DataFileUploadDirectory
         dfud = DataFileUploadDirectory(TEST_CONST.TEST_STREAM_PROC_WATCHED_DIR_PATH_ENCRYPTED,
-                                       TEST_CONST.TEST_CFG_FILE_PATH_ENCRYPTED,
+                                       TEST_CONST.TEST_CFG_FILE_PATH_ENC,
                                        update_secs=UPDATE_SECS,logger=LOGGER)
         #start upload_files_as_added in a separate thread so we can time it out
         upload_thread = ExceptionTrackingThread(
@@ -59,7 +59,7 @@ class TestDataFileStreamProcessorEncyrpted(unittest.TestCase) :
         fp.write_bytes(TEST_CONST.TEST_DATA_FILE_2_PATH.read_bytes())
         time.sleep(5)
         #Use a stream processor to read their data back into memory one time, deliberately failing the first file
-        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH_ENCRYPTED_2,
+        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH_ENC_2,
                                                  TOPIC_NAME,
                                                  output_dir=TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR_RESTART_ENCRYPTED,
                                                  n_threads=RUN_OPT_CONST.N_DEFAULT_DOWNLOAD_THREADS,
@@ -152,7 +152,7 @@ class TestDataFileStreamProcessorEncyrpted(unittest.TestCase) :
         time.sleep(1)
         #recreate and re-run the stream processor, allowing it to successfully process all files this time
         time.sleep(1.0)
-        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH_ENCRYPTED_2,
+        dfsp = DataFileStreamProcessorForTesting(TEST_CONST.TEST_CFG_FILE_PATH_ENC_2,
                                                  TOPIC_NAME,
                                                  output_dir=TEST_CONST.TEST_STREAM_PROCESSOR_OUTPUT_DIR_RESTART_ENCRYPTED,
                                                  n_threads=RUN_OPT_CONST.N_DEFAULT_DOWNLOAD_THREADS,
