+++ docs/source/dev_info/ci_testing.rst
@@ -22,6 +22,7 @@ There are also a few options you can add to ``run_all_tests.py`` if you only wan
 #. Add the ``--no_script_tests`` flag to skip the tests in "test_scripts" entirely, OR
 #. Add the ``--no_kafka`` flag to skip running tests that need to communicate with the Kafka broker. Adding this flag automatically **skips any test methods whose names end with "kafka"**; you will see that they were skipped at the end of the output. Adding this flag also runs tests with some environment variables un-set to ensure that these tests are truly independent of communication with a Kafka broker.
 #. Add the ``--test_regex [regex]`` option to skip any tests whose function names aren't matched to the ``[regex]`` regular expression
+#. Add the ``--local_broker`` flag to run tests using a local Kafka broker running in Docker instead of a third party broker. Including this flag automatically `starts up <https://github.com/openmsi/openmsistream/blob/main/test/start_local_broker.sh#L5>`_ the broker and `creates the topics needed <https://github.com/openmsi/openmsistream/blob/main/test/create_local_testing_topics.sh#L5-L29>`_ before running the tests, and it `stops the broker <https://github.com/openmsi/openmsistream/blob/main/test/stop_local_broker.sh#L5>`_ running when the tests complete. Adding this flag assumes that Docker is installed and that the Docker daemon is running. 
 #. Add the ``--failfast`` flag to stop executing the script early if any individual test(s) fail. Normally all tests are run regardless of how many fail, but including this flag will stop the run as soon as any test fails.
 
 **NOTE:** If you are running the ``run_all_tests.py`` script interactively on a Linux system with ``systemd`` installed, some of the tests will temporarily install code as daemons. You should therefore activate sudo privileges before running the script, with::
@@ -45,13 +46,15 @@ Some of the tests rely on static example data in `test/data <https://github.com/
 External requirements for running tests
 ---------------------------------------
 
-If you would like to run the CI tests using a different cluster than the one set up by the OpenMSI development group, you will need to make several changes.
+All tests can be run using a local Kafka broker by adding the ``--local_broker`` flag to the "run_all_tests" script. Alternatively, tests can use a remote Kafka broker whose bootstrap servers, username, and password are set as the ``KAFKA_TEST_CLUSTER_BOOTSTRAP_SERVERS``, ``KAFKA_TEST_CLUSTER_USERNAME``, and ``KAFKA_TEST_CLUSTER_PASSWORD`` environment variables, respectively. The OpenMSI development group uses a Kafka broker cluster running on `Confluent Cloud <https://confluent.cloud/>`_, and if you get the necessary environment variable values from a developer in the group you won't need to do any extra work.
 
-First, you'll need access to a Kafka cluster. The easiest way to start working with Kafka is to use a cloud-hosted broker by making an account on `the Confluent Cloud website <https://confluent.cloud/>`_. You should create at least one cluster on your account.
+But, if you would like to run the CI tests using a different Kafka broker than the automatically-created local broker or the cluster set up by the OpenMSI development group, you will need to make several changes. First, you'll need access to a Kafka cluster. Options include a self-managed broker running standalone or in Docker (see Kafka documentation for these options), or a broker managed as a service on `the Confluent Cloud website <https://confluent.cloud/>`_.
 
[…]
