### 2023-03-06 #4896 @eminizer
Hi again @SergeyYakubov,

I've just published [v1.3.0 of OpenMSIStream](https://pypi.org/project/openmsistream/1.3.0/), which includes updates from two pull requests, one for each issue you opened. It's really cool to be able to fully test the package using a local broker, and I'm so pleased that the package and its tutorials are now so much more accessible and flexible : )

Thank you again for your excellent suggestions, and please let me know if you run into anything further I should address.

Best wishes,
Maggie
