### 2023-03-07 #4896 @pibion
@eminizer the DOIs suggested by editorialbot are definitely not correct, but could you please update your citation for Pandas to follow their guidance at https://pandas.pydata.org/about/citing.html.  Unfortunately there's no guidance on the Kafka website, I'm working on contacting them to see if they have a recommendation!
