### 2023-03-07 c3999ff
update pandas citation in JOSS paper
