### 2023-03-07 #4896 @pibion
@eminizer a few more citation links for packages that do have DOIs:
- https://numpy.org/citing-numpy/
- https://scipy.org/citing-scipy/
