### 2023-03-07 afb74fa
update numpy, scipy, and pandas citations
