### 2023-03-07 #4896 @eminizer
@pibion the current version on the main branch should now have corrected pandas, numpy, and scipy citations
