### 2023-03-09 #4896 @eminizer
@pibion it looks like I can't get editorialbot to do that for me, but the final official release is v1.3.1, with the archive at DOI 10.5281/zenodo.7713196

Please let me know if you need anything else, and thank you SO MUCH for all of your help!
